cambiando el tipo de dato del id en la clase repository y service de departamento

+++ src/main/java/com/isai/registrerusersjava/app/service/DepartamentService.java
@@ -7,5 +7,5 @@ import java.util.List;
 public interface DepartamentService {
     List<Departament> getAllDepartaments();
 
-    Departament getDepartamentById(Long id);
+    Departament getDepartamentById(Integer id);
 }
